// code/interfaces.ts
export interface AbilityKV
{
    Name: string;
    ID?: number;
    Type?: AbilityTypes;
    Behavior?: AbilityBehavior | Array<AbilityBehavior>;
    OnCastBar?: boolean;
    OnLearnBar?: boolean;
    FightRecapLevel?: 0 | 1 | 2;
    CastRange?: Array<number> | number;
    CastRangeBuffer?: Array<number> | number;
    CastPoint?: Array<number> | number;
    ChannelTime?: Array<number> | number;
    Cooldown?: Array<number> | number;
    Duration?: Array<number> | number;
    SharedCooldown?: string,
    Damage?: Array<number> | number;
    ManaCost?: Array<number> | number;
    ModifierSupportValue?: number;
    ModifierSupportBonus?: number;
    AbilitySpecials?: Array<AbilitySpecialBlock>;
    UnitTargetTeam?: UnitTargetTeam | Array<UnitTargetTeam>;
    CastAnimation?: GameActivity;
    MaxLevel?: number;
    UnitDamageType?: DamageType;
    SpellImmunityType?: SpellImmunityTypes;
    Sound?: string;
    HasScepterUpgrade?: boolean;
    SpellDispellableType?: SpellDispellableTypes;
    HasShardUpgrade?: boolean;
    CastGestureSlot?: AbilityCastGestureSlotValue;
    UnitTargetType?: UnitTargetType | Array<UnitTargetType>;
    AbilityDraftUltScepterAbility?: string;
    GrantedByScepter?: boolean;
    UnitTargetFlags?: UnitTargetFlags | Array<UnitTargetFlags>;
    GrantedByShard?: boolean;
    AbilityDraftUltShardAbility?: string;
    Charges?: Array<number> | number;
    ChargeRestoreTime?: Array<number> | number;
    LinkedAbility?: string;
    AbilityDraftPreAbility?: string;
    ChannelAnimation?: GameActivity;
    IsShardUpgrade?: boolean;
    RequiredLevel?: number;
    LevelsBetweenUpgrades?: number;
    HotKeyOverride?: string;
    DisplayAdditionalHeroes?: boolean;
    TextureName?: string;
    Precache?: Array<PrecacheKV>;
    AnimationPlaybackRate?: Array<number> | number;
    AbilityDraftUltScepterPreAbility?: string;
    Modelscale?: number;
    AssociatedConsumable?: number;
    UnlockMinEffectIndex?: number;
    UnlockMaxEffectIndex?: number;
    EventID?: number;
    ScriptFile?: string,
    AnimationIgnoresModelScale?: boolean;
    IsCastableWhileHidden?: boolean;
}

export enum PropertyType 
{
    Boolean,
    Number,
    String,
    AbilityType,
    AbilityBehaviorOrAbilityBehaviorArray,
    NumberOrNumberArray,
    FightRecapLevel,
    AbilitySpecialBlockArray,
    UnitTargetTeamOrUnitTargetTeamArray,
    GameActivity,
    DamageType,
    SpellImmunityTypes,
    SpellDispellableTypes,
    AbilityCastGestureSlotValue,
    UnitTargetTypeOrUnitTargetTypeArray,
    UnitTargetFlagsOrUnitTargetFlagsArray,
    PrecacheKVArray
};

export const abilityTypeMap: Record<keyof AbilityKV, PropertyType> = 
{
    Name: PropertyType.String,
    ID: PropertyType.Number,
    Type: PropertyType.AbilityType,
    Behavior: PropertyType.AbilityBehaviorOrAbilityBehaviorArray,
    OnCastBar: PropertyType.Boolean,
    OnLearnBar: PropertyType.Boolean,
    FightRecapLevel: PropertyType.FightRecapLevel,
    CastRange: PropertyType.NumberOrNumberArray,
    CastRangeBuffer: PropertyType.NumberOrNumberArray,
    CastPoint: PropertyType.NumberOrNumberArray,
    ChannelTime: PropertyType.NumberOrNumberArray,
    Cooldown: PropertyType.NumberOrNumberArray,
    Duration: PropertyType.NumberOrNumberArray,
    SharedCooldown: PropertyType.String,
    Damage: PropertyType.NumberOrNumberArray,
    ManaCost: PropertyType.NumberOrNumberArray,
    ModifierSupportBonus: PropertyType.Number,
    ModifierSupportValue: PropertyType.Number,
    AbilitySpecials: PropertyType.AbilitySpecialBlockArray,
    UnitTargetTeam: PropertyType.UnitTargetTeamOrUnitTargetTeamArray,
    CastAnimation: PropertyType.GameActivity,
    MaxLevel: PropertyType.Number,
    UnitDamageType: PropertyType.DamageType,
    SpellImmunityType: PropertyType.SpellImmunityTypes,
    Sound: PropertyType.String,
    HasScepterUpgrade: PropertyType.Boolean,
    SpellDispellableType: PropertyType.SpellDispellableTypes,
    HasShardUpgrade: PropertyType.Boolean,
    CastGestureSlot: PropertyType.AbilityCastGestureSlotValue,
    UnitTargetType: PropertyType.UnitTargetTypeOrUnitTargetTypeArray,
    AbilityDraftUltScepterAbility: PropertyType.String,
    Charges: PropertyType.NumberOrNumberArray,
    ChargeRestoreTime: PropertyType.NumberOrNumberArray,
    LinkedAbility: PropertyType.String,
    AbilityDraftPreAbility: PropertyType.String,
    ChannelAnimation: PropertyType.GameActivity,
    IsShardUpgrade: PropertyType.Boolean,
    RequiredLevel: PropertyType.Number,
    LevelsBetweenUpgrades: PropertyType.Number,
    HotKeyOverride: PropertyType.String,
    DisplayAdditionalHeroes: PropertyType.Boolean,
    TextureName: PropertyType.String,
    Precache: PropertyType.PrecacheKVArray,
    AnimationPlaybackRate: PropertyType.NumberOrNumberArray,
    AbilityDraftUltScepterPreAbility: PropertyType.String,
    Modelscale: PropertyType.Number,
    AssociatedConsumable: PropertyType.Number,
    UnlockMinEffectIndex: PropertyType.Number,
    UnlockMaxEffectIndex: PropertyType.Number,
    EventID: PropertyType.Number,
    ScriptFile: PropertyType.String,
    AnimationIgnoresModelScale: PropertyType.Boolean,
    AbilityDraftUltShardAbility: PropertyType.String,
    GrantedByScepter: PropertyType.Boolean,
    GrantedByShard: PropertyType.Boolean,
    IsCastableWhileHidden: PropertyType.Boolean,
    UnitTargetFlags: PropertyType.UnitTargetFlagsOrUnitTargetFlagsArray
}

export enum DifferentlyNamedAbilityKVs
{
    Type = "AbilityType",
    Behavior = "AbilityBehavior",
    OnCastBar = "OnCastbar",
    OnLearnBar = "OnLearnbar",
    CastRange = "AbilityCastRange",
    CastRangeBuffer = "AbilityCastRangeBuffer",
    CastPoint = "AbilityCastPoint",
    ChannelTime = "AbilityChannelTime",
    Cooldown = "AbilityCooldown",
    Duration = "AbilityDuration",
    SharedCooldown = "AbilitySharedCooldown",
    Damage = "AbilityDamage",
    ManaCost = "AbilityManaCost",
    ModifierSupportValue = "AbilityModifierSupportValue",
    ModifierSupportBonus = "AbilityModifierSupportBonus",
    UnitTargetTeam = "AbilityUnitTargetTeam",
    CastAnimation = "AbilityCastAnimation",
    UnitDamageType = "AbilityUnitDamageType",
    Sound = "AbilitySound",
    CastGestureSlot = "AbilityCastGestureSlot",
    UnitTargetType = "AbilityUnitTargetType",
    GrantedByScepter = "IsGrantedByScepter",
    UnitTargetFlags = "AbilityUnitTargetFlags",
    GrantedByShard = "IsGrantedByShard",
    Charges = "AbilityCharges",
    ChargeRestoreTime = "AbilityChargeRestoreTime",
    ChannelAnimation = "AbilityChannelAnimation",
    TextureName = "AbilityTextureName",
    Precache = "precache",
    AbilitySpecials = "AbilitySpecial"
}
// ...
export enum SpellImmunityTypes
{
    ENEMIES_YES = "SPELL_IMMUNITY_ENEMIES_YES",
    ENEMIES_NO = "SPELL_IMMUNITY_ENEMIES_NO",
    ALLIES_NO = "SPELL_IMMUNITY_ALLIES_NO",
    ALLIES_YES = "SPELL_IMMUNITY_ALLIES_YES",
    ALLIES_YES_ENEMIES_NO = "SPELL_IMMUNITY_ALLIES_YES_ENEMIES_NO",
}

export enum SpellDispellableTypes
{
    DISPELLABLE_YES = "SPELL_DISPELLABLE_YES",
    DISPELLABLE_NO = "SPELL_DISPELLABLE_NO",
    DISPELLABLE_YES_STRONG = "SPELL_DISPELLABLE_YES_STRONG"
}

export enum AbilityCastGestureSlotValue
{
    DEFAULT = "DEFAULT",
    ABSOLUTE = "ABSOLUTE"
}

export enum LinkedSpecialBonusOperation
{
    SUBTRACT = "SPECIAL_BONUS_SUBTRACT",
    PERCENTAGE_ADD = "SPECIAL_BONUS_PERCENTAGE_ADD",
    MULTIPLY = "SPECIAL_BONUS_MULTIPLY",
}

export enum PrecacheType
{
    MODEL = "model",
    MODEL_FOLDER = "model_folder",
    PARTICLE = "particle",
    PARTICLE_FOLDER = "particle_folder",
    SOUND = "sound",
    SOUNDFILE = "soundfile"
}

export enum AbilityBehavior
{
    LAST_RESORT_POINT = "DOTA_ABILITY_BEHAVIOR_LAST_RESORT_POINT",
    CAN_SELF_CAST = "DOTA_ABILITY_BEHAVIOR_CAN_SELF_CAST",
    FREE_DRAW_TARGETING = "DOTA_ABILITY_BEHAVIOR_FREE_DRAW_TARGETING",
    IGNORE_SILENCE = "DOTA_ABILITY_BEHAVIOR_IGNORE_SILENCE",
    NONE = "DOTA_ABILITY_BEHAVIOR_NONE",
    SHOW_IN_GUIDES = "DOTA_ABILITY_BEHAVIOR_SHOW_IN_GUIDES",
    SUPPRESS_ASSOCIATED_CONSUMABLE = "DOTA_ABILITY_BEHAVIOR_SUPPRESS_ASSOCIATED_CONSUMABLE",
    UNLOCKED_BY_EFFECT_INDEX = "DOTA_ABILITY_BEHAVIOR_UNLOCKED_BY_EFFECT_INDEX",
    HIDDEN = "DOTA_ABILITY_BEHAVIOR_HIDDEN",
    PASSIVE = "DOTA_ABILITY_BEHAVIOR_PASSIVE",
    NO_TARGET = "DOTA_ABILITY_BEHAVIOR_NO_TARGET",
    UNIT_TARGET = "DOTA_ABILITY_BEHAVIOR_UNIT_TARGET",
    POINT = "DOTA_ABILITY_BEHAVIOR_POINT",
    AOE = "DOTA_ABILITY_BEHAVIOR_AOE",
    NOT_LEARNABLE = "DOTA_ABILITY_BEHAVIOR_NOT_LEARNABLE",
    CHANNELLED = "DOTA_ABILITY_BEHAVIOR_CHANNELLED",
    ITEM = "DOTA_ABILITY_BEHAVIOR_ITEM",
    TOGGLE = "DOTA_ABILITY_BEHAVIOR_TOGGLE",
    DIRECTIONAL = "DOTA_ABILITY_BEHAVIOR_DIRECTIONAL",
    IMMEDIATE = "DOTA_ABILITY_BEHAVIOR_IMMEDIATE",
    AUTOCAST = "DOTA_ABILITY_BEHAVIOR_AUTOCAST",
    OPTIONAL_UNIT_TARGET = "DOTA_ABILITY_BEHAVIOR_OPTIONAL_UNIT_TARGET",
    OPTIONAL_POINT = "DOTA_ABILITY_BEHAVIOR_OPTIONAL_POINT",
    OPTIONAL_NO_TARGET = "DOTA_ABILITY_BEHAVIOR_OPTIONAL_NO_TARGET",
    AURA = "DOTA_ABILITY_BEHAVIOR_AURA",
    ATTACK = "DOTA_ABILITY_BEHAVIOR_ATTACK",
    DONT_RESUME_MOVEMENT = "DOTA_ABILITY_BEHAVIOR_DONT_RESUME_MOVEMENT",
    ROOT_DISABLES = "DOTA_ABILITY_BEHAVIOR_ROOT_DISABLES",
    UNRESTRICTED = "DOTA_ABILITY_BEHAVIOR_UNRESTRICTED",
    IGNORE_PSEUDO_QUEUE = "DOTA_ABILITY_BEHAVIOR_IGNORE_PSEUDO_QUEUE",
    IGNORE_CHANNEL = "DOTA_ABILITY_BEHAVIOR_IGNORE_CHANNEL",
    DONT_CANCEL_MOVEMENT = "DOTA_ABILITY_BEHAVIOR_DONT_CANCEL_MOVEMENT",
    DONT_ALERT_TARGET = "DOTA_ABILITY_BEHAVIOR_DONT_ALERT_TARGET",
    DONT_RESUME_ATTACK = "DOTA_ABILITY_BEHAVIOR_DONT_RESUME_ATTACK",
    NORMAL_WHEN_STOLEN = "DOTA_ABILITY_BEHAVIOR_NORMAL_WHEN_STOLEN",
    IGNORE_BACKSWING = "DOTA_ABILITY_BEHAVIOR_IGNORE_BACKSWING",
    RUNE_TARGET = "DOTA_ABILITY_BEHAVIOR_RUNE_TARGET",
    DONT_CANCEL_CHANNEL = "DOTA_ABILITY_BEHAVIOR_DONT_CANCEL_CHANNEL",
    VECTOR_TARGETING = "DOTA_ABILITY_BEHAVIOR_VECTOR_TARGETING",
}

export enum UnitTargetTeam
{
    NONE = "DOTA_UNIT_TARGET_TEAM_NONE",
    FRIENDLY = "DOTA_UNIT_TARGET_TEAM_FRIENDLY",
    ENEMY = "DOTA_UNIT_TARGET_TEAM_ENEMY",
    BOTH = "DOTA_UNIT_TARGET_TEAM_BOTH",
    CUSTOM = "DOTA_UNIT_TARGET_TEAM_CUSTOM",
}

export enum UnitTargetType
{
    NONE = "DOTA_UNIT_TARGET_NONE",
    HERO = "DOTA_UNIT_TARGET_HERO",
    CREEP = "DOTA_UNIT_TARGET_CREEP",
    BUILDING = "DOTA_UNIT_TARGET_BUILDING",
    COURIER = "DOTA_UNIT_TARGET_COURIER",
    BASIC = "DOTA_UNIT_TARGET_BASIC",
    OTHER = "DOTA_UNIT_TARGET_OTHER",
    ALL = "DOTA_UNIT_TARGET_ALL",
    TREE = "DOTA_UNIT_TARGET_TREE",
    CUSTOM = "DOTA_UNIT_TARGET_CUSTOM",
}

export enum AbilityTypes
{
    BASIC = "ABILITY_TYPE_BASIC",
    ULTIMATE = "ABILITY_TYPE_ULTIMATE",
    ATTRIBUTES = "ABILITY_TYPE_ATTRIBUTES",
    HIDDEN = "ABILITY_TYPE_HIDDEN",
}
// ...
export enum DamageType
{
    NONE = "DAMAGE_TYPE_NONE",
    PHYSICAL = "DAMAGE_TYPE_PHYSICAL",
    MAGICAL = "DAMAGE_TYPE_MAGICAL",
    PURE = "DAMAGE_TYPE_PURE",
    ALL = "DAMAGE_TYPE_ALL",
    HP_REMOVAL = "DAMAGE_TYPE_HP_REMOVAL",
}

export enum UnitTargetFlags
{
    NONE = "DOTA_UNIT_TARGET_FLAG_NONE",
    RANGED_ONLY = "DOTA_UNIT_TARGET_FLAG_RANGED_ONLY",
    MELEE_ONLY = "DOTA_UNIT_TARGET_FLAG_MELEE_ONLY",
    DEAD = "DOTA_UNIT_TARGET_FLAG_DEAD",
    MAGIC_IMMUNE_ENEMIES = "DOTA_UNIT_TARGET_FLAG_MAGIC_IMMUNE_ENEMIES",
    NOT_MAGIC_IMMUNE_ALLIES = "DOTA_UNIT_TARGET_FLAG_NOT_MAGIC_IMMUNE_ALLIES",
    INVULNERABLE = "DOTA_UNIT_TARGET_FLAG_INVULNERABLE",
    FOW_VISIBLE = "DOTA_UNIT_TARGET_FLAG_FOW_VISIBLE",
    NO_INVIS = "DOTA_UNIT_TARGET_FLAG_NO_INVIS",
    NOT_ANCIENTS = "DOTA_UNIT_TARGET_FLAG_NOT_ANCIENTS",
    PLAYER_CONTROLLED = "DOTA_UNIT_TARGET_FLAG_PLAYER_CONTROLLED",
    NOT_DOMINATED = "DOTA_UNIT_TARGET_FLAG_NOT_DOMINATED",
    NOT_SUMMONED = "DOTA_UNIT_TARGET_FLAG_NOT_SUMMONED",
    NOT_ILLUSIONS = "DOTA_UNIT_TARGET_FLAG_NOT_ILLUSIONS",
    NOT_ATTACK_IMMUNE = "DOTA_UNIT_TARGET_FLAG_NOT_ATTACK_IMMUNE",
    MANA_ONLY = "DOTA_UNIT_TARGET_FLAG_MANA_ONLY",
    CHECK_DISABLE_HELP = "DOTA_UNIT_TARGET_FLAG_CHECK_DISABLE_HELP",
    NOT_CREEP_HERO = "DOTA_UNIT_TARGET_FLAG_NOT_CREEP_HERO",
    OUT_OF_WORLD = "DOTA_UNIT_TARGET_FLAG_OUT_OF_WORLD",
    NOT_NIGHTMARED = "DOTA_UNIT_TARGET_FLAG_NOT_NIGHTMARED",
    PREFER_ENEMIES = "DOTA_UNIT_TARGET_FLAG_PREFER_ENEMIES",
    RESPECT_OBSTRUCTIONS = "DOTA_UNIT_TARGET_FLAG_RESPECT_OBSTRUCTIONS",
}